refactor: StructuredTemplateFinder로 이름 변경 및 직접 찾아서 조회할 수 있도록 변경
